+++ bindings/core/tests/secrets_manager.rs
@@ -10,7 +10,7 @@ use tokio::sync::RwLock;
 #[tokio::test]
 async fn generate_ed25519_addresses() -> Result<()> {
     let secret_manager = Arc::new(RwLock::new(SecretManager::try_from_mnemonic(
-        "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river",
+        "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river".to_owned(),
     )?));
 
     let method = SecretManagerMethod::GenerateEd25519Addresses {
@@ -32,7 +32,7 @@ async fn generate_ed25519_addresses() -> Result<()> {
 #[tokio::test]
 async fn generate_evm_addresses() -> Result<()> {
     let secret_manager = Arc::new(RwLock::new(SecretManager::try_from_mnemonic(
-        "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river",
+        "endorse answer radar about source reunion marriage tag sausage weekend frost daring base attack because joke dream slender leisure group reason prepare broken river".to_owned(),
     )?));
 
     let method = SecretManagerMethod::GenerateEvmAddresses {