+++ sdk/tests/wallet/account_recovery.rs
@@ -6,12 +6,7 @@
 use std::time::Duration;
 
 use iota_sdk::{
-    client::{
-        api::GetAddressesOptions,
-        constants::SHIMMER_COIN_TYPE,
-        secret::{mnemonic::MnemonicSecretManager, SecretManager},
-        Client,
-    },
+    client::{api::GetAddressesOptions, constants::SHIMMER_COIN_TYPE, secret::SecretManager, Client},
     wallet::Result,
 };
 
@@ -66,7 +61,7 @@ async fn account_recovery_with_balance_and_empty_addresses() -> Result<()> {
         .finish()
         .await?;
 
-    let secret_manager = SecretManager::Mnemonic(MnemonicSecretManager::try_from_mnemonic(&mnemonic)?);
+    let secret_manager = SecretManager::from(mnemonic.clone());
 
     let addresses = secret_manager
         .generate_ed25519_addresses(
@@ -84,7 +79,7 @@ async fn account_recovery_with_balance_and_empty_addresses() -> Result<()> {
     // Wait for faucet transaction
     tokio::time::sleep(Duration::new(10, 0)).await;
 
-    let wallet = make_wallet(storage_path, Some(&mnemonic), None).await?;
+    let wallet = make_wallet(storage_path, Some(mnemonic), None).await?;
 
     let accounts = wallet.recover_accounts(0, 3, 2, None).await?;